feat: Add non-interactive print mode
This commit introduces a non-interactive 'print' mode, which is triggered by the '--lines <number>' command-line argument. When used, the application will render the moon to the specified number of lines and print it directly to standard output, then exit.

This is useful for MOTD (Message Of The Day) applications, as suggested by user feedback.